+++ src/Styles/home.css
@@ -17,3 +17,8 @@
   justify-content: center;
 }
 
+.button {
+  margin-bottom: 0;
+}
+
+
