--- src/Styles/register.css
@@ -1,5 +1,0 @@
-@import url('./global.css');
-
-body {
-  background-color: #262424;
-}
