+++ src/components/img/HallIcon.js
@@ -1,0 +1,18 @@
+import React from "react";
+import iconsal from "../../assets/icons/iconsal.png";
+import "./styles.css";
+
+
+const HallIcon = () => {
+  return(
+  <div>
+      <img 
+            className="icon" 
+            src={iconsal} 
+            alt="Ícone garçom" 
+        />
+  </div>
+  )
+}
+
+export default HallIcon;
