+++ src/components/img/KitchenIcon.js
@@ -1,0 +1,18 @@
+import React from "react";
+import iconcoz from "../../assets/icons/iconcoz.png";
+import "./styles.css";
+
+
+const KitchenIcon = () => {
+  return (
+      <div>
+          <img 
+          className="icon" 
+          src={iconcoz} 
+          alt="Ícone cozinha" 
+          />
+      </div>
+  )
+}
+
+export default KitchenIcon;
