+++ src/components/img/Logo.js
@@ -1,0 +1,18 @@
+import React from "react";
+import logo from "../../assets/images/logo.png";
+
+import "./styles.css";
+
+const Logo = () => {  
+return (
+    <header className="header">
+          <img 
+              className="responsive center" 
+              src={logo} 
+              alt="Logo Burguer Queen" 
+          />
+    </header>
+)
+}
+
+export default Logo;
